pyproject.toml
[project]
name = "wildtrain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "pycocotools>=2.0.10",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "albumentations>=2.0.8",
    "pillow>=11.3.0",
    "numpy>=2.3.1",
    "python-dotenv>=1.1.1",
    "typer>=0.9.0",
    "dvc>=3.0.0",
    "pre-commit>=4.2.0",
    "spyder-kernels==3.0.*",
]

[project.scripts]
wildtrain = "src.wildtrain.__main__:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
dvc = [
    "dvc[s3]>=3.0.0",
    "dvc[gcs]>=3.0.0",
    "dvc[azure]>=3.0.0",
    "dvc[ssh]>=3.0.0",
]
